feat : create inquiry
- 문의 생성 기능 추가

+++ src/main/java/com/cds/cgv/controller/dto/response/GetTrailerRes.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 @Setter
 public class GetTrailerRes {
     private Long trailerNumber;
-    private int movieNumber;
+    private Long movieNumber;
     private String trailerTitle;
     private String thumbnailLink;
     private String trailerReleasedAt;
